Demographics changes:
(3) Demographics. 
(i) Enable a user to electronically record, change, and access patient
demographic data including preferred language, sex, race, ethnicity, and
date of birth. 
(A) Enable race and ethnicity to be recorded in accordance with the
standard specified in § 170.207(f) and whether a patient declines to
specify race and/or ethnicity. 
(B) Enable preferred language to be recorded in accordance with the
standard specified in § 170.207(g) and whether a patient declines to
specify a preferred language. - See more at:
http://www.open-emr.org/wiki/index.php/Demographics_(MU2)#sthash.uTvBfKjH.dpuf